+++ app/prayers/[id]/_components/index/PrayerRoomBody.tsx
@@ -1,4 +1,4 @@
-import PrayerTitleList from "../../../../../src/domain/prayerTitles/components/prayerTitleList";
+import PrayerTitleList from "../../../../rooms/_components/[id]/prayerTitleList";
 
 export default function PrayerRoomBody(): JSX.Element {
   return <PrayerTitleList />;
