+++ app/prayers/[id]/update.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useMemo, useEffect, useRef } from "react";
 import { StyleSheet, SafeAreaView, BackHandler } from "react-native";
-import { useSelectedPrayerTitleStore } from "../../../src/domain/prayerTitles/stores/useSelectedPrayerTitleStore";
+import { useSelectedPrayerTitleStore } from "../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
 import { useSelectedRoomStore } from "../../../src/domain/rooms/stores/useSelectedRoomStore";
 import { usePrayerCreationStore } from "../../../src/domain/prayerCreation/stores/usePrayerCreationStore";
 import { router } from "expo-router";
