+++ src/domain/prayerRead/components/PrayerReadBody.tsx
@@ -5,7 +5,7 @@ import { PrayerContent } from "../types/response/prayerContent";
 import TitleCard from "./TitleCard";
 import PrayerCardList from "./PrayerCardList";
 import EmptyState from "./EmptyState";
-import { useSelectedPrayerTitleStore } from "../../prayerTitles/stores/useSelectedPrayerTitleStore";
+import { useSelectedPrayerTitleStore } from "../../prayers/stores/useSelectedPrayerTitleStore";
 import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
 import { usePrayerContentsQuery } from "../hooks/queries/usePrayerReadQuery";
