--- src/domain/prayerTitles/services/prayerTitleService.ts
@@ -1,18 +1,0 @@
-import apiService from "../../../common/apis/apiService";
-import api, { ApiResponse } from "../../../common/apis/api";
-import { PrayerTitle } from "../types/dto/response/prayerTitle";
-
-const prayerTitleService = {
-  fetchPrayerTitles: async (roomId: number | null, after: string = "0") => {
-    const response = await apiService.get<{ prayerTitles: PrayerTitle[] }>(
-      "/prayers",
-      {
-        roomId,
-        after,
-      }
-    );
-    return response.data.prayerTitles;
-  },
-};
-
-export default prayerTitleService;
