+++ src/domain/prayers/services/prayerServices.ts
@@ -3,6 +3,8 @@ import { ApiResponse } from "../../../common/apis/api";
 import type { PrayerCreationItem } from "../types/PrayerCreationItem";
 import { CreatePrayerRequest } from "../types/request/createPrayerRequest";
 import { MessageResponse } from "../../../common/types/messageResponse";
+import { PrayerTitle } from "../types/prayerTitle";
+import { FetchPrayerTitlesResponse } from "../types/response/fetchPrayerTitlesResponse";
 
 export const prayerService = {
   // 기도 제목 작성
@@ -19,5 +21,16 @@ export const prayerService = {
       } as CreatePrayerRequest);
     console.log("API response=", response.message);
     return response;
+  },
+  // 기도 제목 무한 스크롤 조회
+  fetchTitles: async (roomId: number | null, after: string = "0") => {
+    const response = await apiService.get<FetchPrayerTitlesResponse>(
+      "/prayers",
+      {
+        roomId,
+        after,
+      }
+    );
+    return response.data.prayerTitles;
   },
 };
