+++ src/domain/prayers/types/params/fetchPrayerTitlesParams.ts
@@ -1,0 +1,4 @@
+export interface FetchPrayerTitlesParams {
+  roomId: number | null;
+  after: string;
+}
