+++ src/domain/prayers/types/response/fetchPrayerTitlesResponse.ts
@@ -1,0 +1,5 @@
+import { PrayerTitle } from "../prayerTitle";
+
+export interface FetchPrayerTitlesResponse {
+  prayerTitles: PrayerTitle[];
+}
